refactoring NotificationTimelineDao
