+++ src/main/java/it/pagopa/pn/datavault/middleware/db/NotificationTimelineDao.java
@@ -1,30 +1,30 @@
 package it.pagopa.pn.datavault.middleware.db;
 
+import it.pagopa.pn.commons.db.BaseDAO;
 import it.pagopa.pn.datavault.config.PnDatavaultConfig;
+import it.pagopa.pn.datavault.generated.openapi.server.v1.dto.ConfidentialTimelineElementId;
 import it.pagopa.pn.datavault.middleware.db.entities.NotificationTimelineEntity;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import software.amazon.awssdk.enhanced.dynamodb.DynamoDbAsyncTable;
+import reactor.util.function.Tuples;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedAsyncClient;
-import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
-import software.amazon.awssdk.enhanced.dynamodb.model.Page;
 import software.amazon.awssdk.enhanced.dynamodb.model.QueryConditional;
-import software.amazon.awssdk.enhanced.dynamodb.model.QueryEnhancedRequest;
+import software.amazon.awssdk.services.dynamodb.DynamoDbAsyncClient;
+
+import static it.pagopa.pn.datavault.middleware.db.entities.NotificationTimelineEntity.ADDRESS_PREFIX;
 
 @Repository
 @Slf4j
-public class NotificationTimelineDao extends BaseDao {
-
-    DynamoDbAsyncTable<NotificationTimelineEntity> timelineTable;
-    DynamoDbEnhancedAsyncClient dynamoDbEnhancedAsyncClient;
+public class NotificationTimelineDao extends BaseDAO<NotificationTimelineEntity> {
 
 
-    public NotificationTimelineDao(DynamoDbEnhancedAsyncClient dynamoDbEnhancedAsyncClient,
-                                   PnDatavaultConfig pnDatavaultConfig) {
-        this.timelineTable = dynamoDbEnhancedAsyncClient.table(pnDatavaultConfig.getDynamodbTableName(), TableSchema.fromBean(NotificationTimelineEntity.class));
-        this.dynamoDbEnhancedAsyncClient = dynamoDbEnhancedAsyncClient;
+    protected NotificationTimelineDao(DynamoDbEnhancedAsyncClient dynamoDbEnhancedAsyncClient,
+                                      DynamoDbAsyncClient dynamoDbAsyncClient,
+                                      PnDatavaultConfig pnDatavaultConfig
+                                       ) {
+        super(dynamoDbEnhancedAsyncClient, dynamoDbAsyncClient, pnDatavaultConfig.getDynamodbTableName(), NotificationTimelineEntity.class);
     }
 
     /**
@@ -36,31 +36,37 @@ public class NotificationTimelineDao extends BaseDao {
     {
         log.debug("updateNotification timeline internalid:{} timelineelementid:{}",entity.getInternalId(), entity.getTimelineElementId());
 
-        return Mono.fromFuture(timelineTable.updateItem(entity));
+        return Mono.fromFuture(update(entity));
     }
 
     public Mono<NotificationTimelineEntity> getNotificationTimelineByIunAndTimelineElementId(String iun, String timelineElementId)
     {
         log.debug("getNotificationTimelineByIunAndTimelineElementId timeline internalid:{} timelineelementid:{}", iun, timelineElementId);
 
-        return Mono.fromFuture(timelineTable.getItem(new NotificationTimelineEntity(iun, timelineElementId)));
+        return Mono.fromFuture(get(getIunWhitPrefix(iun), timelineElementId));
     }
 
     public Flux<NotificationTimelineEntity> getNotificationTimelineByIun(String iun) {
         log.debug("getNotificationTimelineByIun timelines list-by-id internalid:{}", iun);
 
-        NotificationTimelineEntity ne = new NotificationTimelineEntity(iun, "");
-        QueryConditional queryConditional = QueryConditional.keyEqualTo(getKeyBuild(ne.getPk()));
-
-        QueryEnhancedRequest qeRequest = QueryEnhancedRequest
-                .builder()
-                .queryConditional(queryConditional)
-                .scanIndexForward(true)
-                .build();
+        QueryConditional queryConditional = QueryConditional.keyEqualTo(keyBuild(getIunWhitPrefix(iun), null));
 
         // viene volutamente ignorata la gestione della paginazione, che per ora non serve.
         // si suppone infatti che la lista degli indirizzi non sia troppo lunga e quindi non vada a sforare il limite di 1MB di paginazione
-        return Flux.from(timelineTable.query(qeRequest)
-                .flatMapIterable(Page::items));
+        return Flux.from(getByFilter(queryConditional, null, null, null));
+    }
+
+    public Flux<NotificationTimelineEntity> getNotificationTimelines(Flux<ConfidentialTimelineElementId> confidentialTimelineElementId) {
+
+        log.debug("getNotification timelines confidentialTimelineElementId:{}", confidentialTimelineElementId);
+
+        return confidentialTimelineElementId
+                .map(id -> Tuples.of(getIunWhitPrefix(id.getIun()), id.getTimelineElementId()))
+                .collectList()
+                .flatMapMany(this::batchGetItem);
+    }
+
+    private String getIunWhitPrefix (String iun){
+        return ADDRESS_PREFIX + iun;
     }
 }
